Separate logic from UnitTableDataSource into UnitService
- This fixes problem of sharing the same data between the DB Table and the team-builder
- Now each module/component can receive a separate instance of UnitService which it can use to get all units
