+++ src/app/core/services/unit.service.ts
@@ -1,86 +1,19 @@
-import { DataSource } from '@angular/cdk/collections';
-import { MatPaginator } from '@angular/material/paginator';
-import { MatSort } from '@angular/material/sort';
-import { MatTableDataSource } from '@angular/material/table';
-import { map } from 'rxjs/operators';
+import * as rumble from '../../shared/models/rumble';
+import rumbleData from '../../../assets/data/rumble.json';
+import { UnitDetails } from '../../shared/models/unit-details';
 import { Injectable } from '@angular/core';
-import { Observable, of as observableOf, merge } from 'rxjs';
-import { BehaviorSubject } from 'rxjs';
-import * as rumble from '../../../shared/models/rumble';
-import rumbleData from '../../../../assets/data/rumble.json';
 
 declare const window: any;
 
-export class UnitDetails {
-  complete: boolean;
-  id: number;
-  isBaseForm: boolean;
-  name: string;
-  baseHp: number;
-  baseAtk: number;
-  baseRcv: number;
-  baseDef: number;
-  baseSpd: number;
-  style: string;
-  type: string;
-  class1: string;
-  class2: string;
-  festAbility: string;
-  festSpecial: string;
-  festCooldown: string;
-  festAttackPattern: string;
-  festAttackTarget: string;
+// dont declare service as singleton (i.e., providedIn: 'root')
+// so that each client of the service can get an instance based on its needs
+// (eg, instance per module or instance per component)
+@Injectable()
+class UserService {
 
-/*
-  private computeDefPercentage(): number {
-    return 9.6671298720642E-17 * this.baseDef ** 6
-            - 2.54940335570173E-13 * this.baseDef ** 5
-            + 0.000000000265685114974299 * this.baseDef ** 4
-            - 0.000000135059490498531 * this.baseDef ** 3
-            + 0.000030881486606793 * this.baseDef ** 2
-            - 0.000309272395298876 * this.baseDef
-            + 0.00126339332598491 * 100;
-  }
-
-  public getId(): number { return this.id; }
-  public getName(): string { return this.name; }
-  public getBaseHp(): number { return this.baseHp; }
-  public getBaseAtk(): number { return this.baseAtk; }
-  public getBaseRcv(): number { return this.baseRcv; }
-  public getBaseDef(): number { return Math.round((this.computeDefPercentage() + Number.EPSILON) * 100) / 100; }
-  public getBaseSpd(): number { return this.baseSpd; }
-  public getstyle(): string { return this.style; }
-  public getType(): string { return '<span class="badge text-monospace badge-pill ' + this.type + '">' + this.type + '</span>'; }
-  public getFestAbility(): string { return this.encodeText(this.festAbility[4].description); }
-  public getFestSpecial(): string { return this.encodeText(this.festSpecial[9].description); }
-  public getFestAttackPattern(): string { return this.festAttackPattern; }
-  public getFestAttackTarget(): string { return this.festAttackTarget; }
-
-  private encodeText(s: string): string
-  {
-    return s.replace(/\[([A-Z]+)\]/gi,
-                     '<span class="badge text-monospace badge-pill $1">$1</span>');
-  }
-  */
-}
-
-/**
- * Data source for the UnitTable view. This class should
- * encapsulate all logic for fetching and manipulating the displayed data
- * (including sorting, pagination, and filtering).
- */
-@Injectable({
-   providedIn: 'root'
- })
-export class UnitTableDataSource extends MatTableDataSource<rumble.Unit> {
-  database: rumble.Unit[] = [];
-
-  filterChainSubject: BehaviorSubject<((unit: rumble.Unit) => boolean) []>
-   = new BehaviorSubject<((unit: rumble.Unit) => boolean) []>([unit => !unit.isBaseForm]);
+  private rumbleUnits: rumble.Unit[] = [];
 
   constructor() {
-    super();
-
     let unitDetail: UnitDetails;
     const unitDetails: UnitDetails[] = [];
     let unit: rumble.Unit;
@@ -190,14 +123,11 @@ export class UnitTableDataSource extends MatTableDataSource<rumble.Unit> {
       unit.lvl10Special = (unit.special[9].effects as rumble.Effect[]);
       unit.lvl10Cooldown = unit.special[9].cooldown;
       unit.thumbnailUrl = window.Utils.getThumbnailUrl(Math.floor(unit.id)).replace('..', 'https://optc-db.github.io/');
-      this.database.push(unit);
+      this.rumbleUnits.push(unit);
     }
-
-    this.filterData();
-
   }
 
-  denormalizeEffects(levels: rumble.Ability[]|rumble.Special[]): void {
+  private denormalizeEffects(levels: rumble.Ability[]|rumble.Special[]): void {
     const lastEffect: rumble.Effect[] = [];
     let mergedEffect: rumble.Effect[] = [];
 
@@ -215,10 +145,10 @@ export class UnitTableDataSource extends MatTableDataSource<rumble.Unit> {
     });
   }
 
-  filterData(): void {
-    let tmpData: rumble.Unit[] = this.database;
-    this.filterChainSubject.value.forEach(unitFilter => tmpData = tmpData.filter(unitFilter));
-    this.data = tmpData;
-    this._updateChangeSubscription();
-  }
+  // we can make a shallow copy of the array if we are worried
+  // about clients directly modifying the array,
+  // but thats something they shouldnt do in the first place
+  public getUnits = () => this.rumbleUnits;
 }
+
+export default UserService;
