+++ src/app/database/components/unit-table/unit-table.component.ts
@@ -2,7 +2,7 @@ import { AfterViewInit, Component, OnInit, ViewChild } from '@angular/core';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { MatTable } from '@angular/material/table';
-import { UnitTableDataSource } from './unit-table-datasource';
+import { UnitTableDataSource } from '../../services/unit-table-datasource';
 import { Unit } from '../../../shared/models/rumble';
 
 @Component({
