+++ src/app/database/services/unit-table-datasource.ts
@@ -1,0 +1,33 @@
+import { MatTableDataSource } from '@angular/material/table';
+import { BehaviorSubject } from 'rxjs';
+import * as rumble from '../../shared/models/rumble';
+import UnitService from '../../core/services/unit.service';
+import { Injectable } from '@angular/core';
+
+/**
+ * Data source for the UnitTable view. This class should
+ * encapsulate all logic for fetching and manipulating the displayed data
+ * (including sorting, pagination, and filtering).
+ */
+@Injectable()
+export class UnitTableDataSource extends MatTableDataSource<rumble.Unit> {
+  database: rumble.Unit[] = [];
+
+  filterChainSubject: BehaviorSubject<((unit: rumble.Unit) => boolean) []>
+   = new BehaviorSubject<((unit: rumble.Unit) => boolean) []>([unit => !unit.isBaseForm]);
+
+  constructor(unitService: UnitService) {
+    super();
+    for (const unit of unitService.getUnits()) {
+      this.database.push(unit);
+    }
+    this.filterData();
+  }
+
+  filterData(): void {
+    let tmpData: rumble.Unit[] = this.database;
+    this.filterChainSubject.value.forEach(unitFilter => tmpData = tmpData.filter(unitFilter));
+    this.data = tmpData;
+    this._updateChangeSubscription();
+  }
+}
